+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.1.4
 
- * Upgrade Kryo to 4.0.1
+ * Upgrade Kryo to 4.0.1 #121
+ * Add option to disable parallel repository indexing on startup #121
 
 ## 1.1.3